requirements.txt
black==25.1.0
click==8.1.8
flake8==7.2.0
ghidriff==0.8.0
jpype1==1.5.2
load-dotenv==0.1.0
Markdown==3.7
mccabe==0.7.0
mdutils==1.6.0
mypy-extensions==1.0.0
narwhals==1.32.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
plotly==6.0.1
PuLP==3.0.2
pycodestyle==2.13.0
pyflakes==3.3.2
pyghidra==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
shortuuid==1.0.13
six==1.17.0
tzdata==2025.2
PyYAML
